Updated simulation names of performance tests to include Filter or UseCase (e.g. SamlFilterSimulation)
