+++ repose-aggregator/tests/performance-tests/test_vars/filters/herp.yml
@@ -17,7 +17,7 @@
   },
   gatling: {
     test: {
-      name: "HerpSimulation",
+      name: "HerpFilterSimulation",
       package: "filters.herp"
     }
   }
