+++ repose-aggregator/tests/performance-tests/test_vars/filters/saml.yml
@@ -38,7 +38,7 @@
   },
   gatling: {
     test: {
-      name: "SamlSimulation",
+      name: "SamlFilterSimulation",
       package: "filters.saml",
       config: "{{ dir_prefix }}/saml_simulation.conf",
       params: {
