+++ repose-aggregator/tests/performance-tests/test_vars/filters/valkyrie.yml
@@ -24,7 +24,7 @@
   },
   gatling: {
     test: {
-      name: "ValkyrieSimulation",
+      name: "ValkyrieFilterSimulation",
       package: "filters.valkyrie"
     }
   }
